Adição da tabela que armazena as localizações do veiculo, bem como o repository para pesquisa, adição do endpoint para buscar essas informações, ajustes nos DTO's, entities, mappings entre outros objetos para que o processo funcione corretamente.

+++ src/GeoTruck.Services.Domain/Entities/VehicleLocation.cs
@@ -6,15 +6,15 @@ public class VehicleLocation : Entity<VehicleLocation>
 {
     public int VehicleId { get; protected set; }
     public virtual Vehicle Vehicle { get; protected set; }
-    public double Latitude { get; protected set; }
-    public double Longitude { get; protected set; }
+    public decimal Latitude { get; protected set; }
+    public decimal Longitude { get; protected set; }
     public long PositionId { get; protected set; }
     public DateTime Date { get; protected set; }
     public DateTime DateUTC { get; protected set; }
 
     protected VehicleLocation() { }
 
-    public VehicleLocation(Vehicle vehicle, double latitude, double longitude, long positionId, DateTime date, DateTime dateUTC)
+    public VehicleLocation(Vehicle vehicle, decimal latitude, decimal longitude, long positionId, DateTime date, DateTime dateUTC)
     {
         Vehicle = vehicle;
         VehicleId = vehicle.Id;
@@ -25,7 +25,7 @@ public class VehicleLocation : Entity<VehicleLocation>
         DateUTC = dateUTC;
     }
 
-    public static VehicleLocation Create(Vehicle vehicle, double latitude, double longitude, long positionId, DateTime date, DateTime dateUTC)
+    public static VehicleLocation Create(Vehicle vehicle, decimal latitude, decimal longitude, long positionId, DateTime date, DateTime dateUTC)
     {
         return new VehicleLocation(vehicle, latitude, longitude, positionId, date, dateUTC);
     }
